requirements.txt
flask==3.0.3
flask-cors==4.0.1
sqlalchemy==2.0.32
pillow==10.4.0
numpy==1.26.4
tensorflow==2.17.0
keras==3.4.1

